chore: commit pending changes across mycelix SDK and symthaea
Mycelix: governance SDK client improvements (bridge, threshold-signing,
voting types), food integration expansion, DNA size audit update.

Symthaea: cognitive loop enrichment (cycle, stats, types), database
client fixes (lance, sqlite), psych-bench tuning (12 benchmarks),
sub-crate fixes (dynamics, factor-graph, flight, observability),
memory/visualization/voice minor fixes, integration test expansion.

+++ symthaea/crates/symthaea-psych-bench/src/benchmarks/cogbench/bart.rs
@@ -129,7 +129,7 @@ impl BartBenchmark {
             let fep_pump_prob = action_result.action_probabilities
                 .first()
                 .copied()
-                .unwrap_or(0.5) as f64;
+                .unwrap_or(0.5);
 
             // FEP modulates target by +/- 20%
             let fep_adjustment = (fep_pump_prob - 0.5) * 0.4;
